requirements.txt
google.generativeai 
transformers 
torch 
sacremoses
pinecone-client[grpc] 
sentence-transformers
python-dotenv